refactor(modifiers): rename ModifiedAbilityModifiers to ModifiersList
